Update patch set 27
Patch Set 27:

We'll need to add python-sushy to fedora. Currently is only in RDO, but as rsdclient is to be added to fedora (it as a openstackclient module), it's needed in fedora too. I would not block this package in RDO because of this but in parallel, could you create a package review bugzilla for python-sushy?

Patch-set: 27